# lxdapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "魔方财务系统专用LXD容器管理后端API服务",
        "title": "魔方财务-LXD对接API",
        "termsOfService": "https://github.com/xkatld/zjmf-lxd-server",
        "contact": {
            "name": "API支持",
            "url": "https://github.com/xkatld/zjmf-lxd-server",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE"
        },
        "version": "1.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/console/create-token": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建用于WebSocket控制台连接的安全令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Web控制台"
                ],
                "summary": "创建控制台访问令牌",
                "parameters": [
                    {
                        "description": "令牌参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/console/sessions/{sessionId}": {
            "delete": {
                "tags": [
                    "Console"
                ],
                "summary": "关闭会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "关闭结果",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/console/status": {
            "get": {
                "tags": [
                    "Console"
                ],
                "summary": "获取控制台状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "hostname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "状态信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/console/ws": {
            "get": {
                "tags": [
                    "Console"
                ],
                "summary": "WebSocket控制台连接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "hostname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/console/ws-token": {
            "get": {
                "tags": [
                    "Web控制台"
                ],
                "summary": "基于令牌的WebSocket控制台",
                "parameters": [
                    {
                        "type": "string",
                        "description": "访问令牌",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/console": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Web控制台"
                ],
                "summary": "控制台Web页面",
                "parameters": [
                    {
                        "type": "string",
                        "description": "访问令牌",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "models.CreateTokenRequest": {
            "type": "object",
            "required": [
                "hostname"
            ],
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "hostname": {
                    "type": "string"
                },
                "server_ip": {
                    "type": "string"
                },
                "service_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "在header中添加apikey字段，值为配置文件中的api_hash",
            "type": "apiKey",
            "name": "apikey",
            "in": "header"
        }
    }
}